% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_format.R
\name{get_format}
\alias{get_format}
\title{Gets the format attributes of the variables of a list or data frame}
\usage{
get_format(df)
}
\arguments{
\item{df}{a list or data frame}
}
\description{
Gets the format attributes of the variables of a list or data frame
}
\examples{
## retrieving the format of an Stata version of the iris data set:
library(haven)
path <- system.file("examples", "iris.dta", package = "haven")
iris <- read_dta(path)
str(iris)
get_format(iris)

}
\author{
Marc Choisy
}
