% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_format.R
\name{trim_format}
\alias{trim_format}
\title{Trims the format attributes of the variables of a list or data frame}
\usage{
trim_format(df)
}
\arguments{
\item{df}{a list or data frame}
}
\description{
Trims the format attributes of the variables of a list or data frame
}
\examples{
## retrieving the format of an Stata version of the iris data set:
library(haven)
path <- system.file("examples", "iris.dta", package = "haven")
iris <- read_dta(path)
str(iris)
iris <- trim_format(iris)
str(iris)

}
\author{
Marc Choisy
}
