% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_def.R
\name{var_def}
\alias{var_def}
\title{Returns the definitions of the variables}
\usage{
var_def(data, ...)
}
\arguments{
\item{df}{a list or data frame}
}
\description{
Returns the definitions of the variables
}
\examples{
## retrieving the format of an Stata version of the iris data set:
library(haven)
path <- system.file("examples", "iris.dta", package = "haven")
iris <- read_dta(path)
str(iris)
var_def(iris)
var_def(iris, sepallength)
var_def(iris, sepallength, petallength)
var_def(iris[c("sepallength", "petallength")])
var_def(iris[, c("sepallength", "petallength")])
all_about_length <- grep("length", names(iris), value = TRUE)
var_def(iris[all_about_length])

}
\author{
Marc Choisy
}
